requirements.txt
dask-ms
numpy
numba
numpydoc
scipy
SharedArray
matplotlib
